refactor: :sparkles: History Model 구독 컴포넌트 변경
- History Model을 캘린더 body나 footer 가 아닌 상위의 캘린더 컴포넌트에서 구독하도록 설정
- body, footer 에서는 history 데이터를 props로 전달받음
