+++ client/src/views/components/calendar/Footer/Footer.js
@@ -2,25 +2,27 @@ import Component from '@lib/Component';
 import './Footer.scss';
 
 class CalendarFooter extends Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
 
     this.init();
   }
 
   render() {
+    const { total } = this.props;
+
     const $calendarFooter = document.createElement('div');
     $calendarFooter.className = 'calendar-footer';
 
     $calendarFooter.innerHTML = `
       <div class="calendar-footer__left">
         <div class="calendar-footer__left--pr">
-          총 수입 <span>1,822,480</span>
+          총 수입 <span>${total.income}</span>
         </div>
-        <div>총 지출 <span>834,640</span></div>
+        <div>총 지출 <span>${total.expenses}</span></div>
       </div>
       <div class="calendar-footer__right">
-        총계 <span>987,840</span>
+        총계 <span>${total.earning}</span>
       </div>
     `;
